fix: Minor refactor
